use mybaties genator

+++ src/main/java/com/demo/community/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Long id, Model model){
+    public String question(@PathVariable(name = "id") Integer id, Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("question",questionDTO);
         return "question";
